Add progress status while downloading a file

+++ lib/src/web/web_helper.dart
@@ -1,16 +1,22 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+import 'package:flutter_cache_manager/src/result/file_response.dart';
 import 'package:flutter_cache_manager/src/storage/cache_object.dart';
 import 'package:flutter_cache_manager/src/cache_store.dart';
 import 'package:flutter_cache_manager/src/web/file_fetcher.dart';
-import 'package:flutter_cache_manager/src/file_info.dart';
+import 'package:flutter_cache_manager/src/result/file_info.dart';
 import 'package:pedantic/pedantic.dart';
+import 'package:rxdart/rxdart.dart';
 import 'package:uuid/uuid.dart';
 
 ///Flutter Cache Manager
 ///Copyright (c) 2019 Rene Floor
 ///Released under MIT License.
+
+const statusCodesNewFile = [HttpStatus.ok, HttpStatus.accepted];
+const statusCodesFileNotChanged = [HttpStatus.notModified];
 
 class WebHelper {
   WebHelper(this._store, FileService fileFetcher)
@@ -19,82 +25,85 @@ class WebHelper {
 
   final CacheStore _store;
   final FileService _fileFetcher;
-  final Map<String, Future<FileInfo>> _memCache;
+  final Map<String, BehaviorSubject<FileResponse>> _memCache;
 
   ///Download the file from the url
-  Future<FileInfo> downloadFile(String url,
-      {Map<String, String> authHeaders, bool ignoreMemCache = false}) async {
+  Stream<FileResponse> downloadFile(String url,
+      {Map<String, String> authHeaders, bool ignoreMemCache = false}) {
     if (!_memCache.containsKey(url) || ignoreMemCache) {
-      var completer = Completer<FileInfo>();
+      var subject = BehaviorSubject<FileResponse>();
+      _memCache[url] = subject;
+
       unawaited(() async {
         try {
-          final cacheObject =
-              await _downloadRemoteFile(url, authHeaders: authHeaders);
-          completer.complete(cacheObject);
+           await for(var result in _updateFile(url, authHeaders: authHeaders)) {
+             subject.add(result);
+           }
         } catch (e, stackTrace) {
-          completer.completeError(e);
+          subject.addError(e, stackTrace);
         } finally {
-          unawaited(_memCache.remove(url));
+          await subject.close();
+          _memCache.remove(url);
         }
       }());
-      _memCache[url] = completer.future;
     }
-    return _memCache[url];
+    return _memCache[url].stream;
   }
 
   ///Download the file from the url
-  Future<FileInfo> _downloadRemoteFile(String url,
-      {Map<String, String> authHeaders}) async {
+  Stream<FileResponse> _updateFile(String url,
+      {Map<String, String> authHeaders}) async* {
     var cacheObject = await _store.retrieveCacheData(url);
     cacheObject ??= CacheObject(url);
+    final response = await _download(cacheObject, authHeaders);
+    yield* _manageResponse(cacheObject, response);
 
+    final file = (await _store.fileDir).childFile(cacheObject.relativePath);
+    yield FileInfo(file, FileSource.Online, cacheObject.validTill, url);
+  }
+
+  Future<FileFetcherResponse> _download(
+      CacheObject cacheObject, Map<String, String> authHeaders) {
     final headers = <String, String>{};
     if (authHeaders != null) {
       headers.addAll(authHeaders);
     }
 
     if (cacheObject.eTag != null) {
-      headers['If-None-Match'] = cacheObject.eTag;
+      headers[HttpHeaders.ifNoneMatchHeader] = cacheObject.eTag;
     }
 
-    final response = await _fileFetcher.get(url, headers: headers);
-    final success = await _handleHttpResponse(response, cacheObject);
-    if (!success) {
+    return _fileFetcher.get(cacheObject.url, headers: headers);
+  }
+
+  Stream<DownloadProgress> _manageResponse(
+      CacheObject cacheObject, FileFetcherResponse response) async* {
+    final hasNewFile = statusCodesNewFile.contains(response.statusCode);
+    final keepOldFile = statusCodesFileNotChanged.contains(response.statusCode);
+    if (!hasNewFile && !keepOldFile) {
       throw HttpExceptionWithStatus(
         response.statusCode,
         'Invalid statusCode: ${response?.statusCode}',
-        uri: Uri.parse(url),
+        uri: Uri.parse(cacheObject.url),
       );
     }
 
-    unawaited(_store.putFile(cacheObject));
-
-    final file = (await _store.fileDir).childFile(cacheObject.relativePath);
-    return FileInfo(file, FileSource.Online, cacheObject.validTill, url);
-  }
-
-  Future<bool> _handleHttpResponse(
-      FileFetcherResponse response, CacheObject cacheObject) async {
-    if (response.statusCode == 200 || response.statusCode == 201) {
-      final basePath = await _store.fileDir;
-      _setDataFromHeaders(cacheObject, response);
-      final file = basePath.childFile(cacheObject.relativePath);
-      final folder = file.parent;
-      if (!(await folder.exists())) {
-        folder.createSync(recursive: true);
+    final oldCacheFile = cacheObject.relativePath;
+    var newCacheFile = cacheObject.relativePath;
+    _setDataFromHeaders(cacheObject, response);
+    if (statusCodesNewFile.contains(response.statusCode)) {
+      await for (var progress in _saveFile(cacheObject, response)) {
+        yield DownloadProgress(
+            cacheObject.url, response.contentLength, progress);
       }
-
-      final sink = file.openWrite();
-      await sink.addStream(response.content);
-      await sink.close();
-
-      return true;
+      newCacheFile = cacheObject.relativePath;
     }
-    if (response.statusCode == 304) {
-      _setDataFromHeaders(cacheObject, response);
-      return true;
-    }
-    return false;
+
+    unawaited(_store.putFile(cacheObject).then((_) {
+      if (newCacheFile != oldCacheFile) {
+        _removeOldFile(oldCacheFile);
+      }
+    }));
   }
 
   void _setDataFromHeaders(
@@ -112,7 +121,43 @@ class WebHelper {
     cacheObject.relativePath ??= '${Uuid().v1()}$fileExtension';
   }
 
+  Stream<int> _saveFile(CacheObject cacheObject, FileFetcherResponse response) {
+    var receivedBytesResultController = StreamController<int>();
+    unawaited(_saveFileAndPostUpdates(
+      receivedBytesResultController,
+      cacheObject,
+      response,
+    ));
+    return receivedBytesResultController.stream;
+  }
+
+  Future _saveFileAndPostUpdates(
+      StreamController<int> receivedBytesResultController,
+      CacheObject cacheObject,
+      FileFetcherResponse response) async {
+    final basePath = await _store.fileDir;
+
+    final file = basePath.childFile(cacheObject.relativePath);
+    final folder = file.parent;
+    if (!(await folder.exists())) {
+      folder.createSync(recursive: true);
+    }
+    try {
+      var receivedBytes = 0;
+      final sink = file.openWrite();
+      await response.content.map((s) {
+        receivedBytes += s.length;
+        receivedBytesResultController.add(receivedBytes);
+        return s;
+      }).pipe(sink);
+    }catch(e, stacktrace){
+      receivedBytesResultController.addError(e, stacktrace);
+    }
+    await receivedBytesResultController.close();
+  }
+
   Future<void> _removeOldFile(String relativePath) async {
+    if (relativePath == null) return;
     final file = (await _store.fileDir).childFile(relativePath);
     if (await file.exists()) {
       await file.delete();
